pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pynoddgcs"
version = "0.1.0"
description = "Utilities for connecting to a gcs bucket and uploading files"
authors = [
    { name = "CScott Brown" },
]
dependencies = [
  "google-cloud-storage",
  "google-auth",
  "requests"
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
]
